delete one nilai pertemuan done

+++ backend/src/controllers/tahunAjaran-controller.js
@@ -15,16 +15,28 @@ export const addAjaran = async (req, res, next) => {
       throw new ResponseError(400, "Format Tahun ajaran tidak sesuai");
     }
 
-    await TahunAjaran.updateMany({ status: false });
+    const [first, last] = ajaran.split("/");
 
-    const newAjaran = new TahunAjaran({ ajaran });
+    const firstYear = parseInt(first);
+    const lastYear = parseInt(last);
 
-    await newAjaran.save();
+    if (lastYear - firstYear === 1) {
+      await TahunAjaran.updateMany({ status: false });
 
-    res.status(200).json({
-      success: true,
-      message: "Berhasil menambah Tahun Ajaran.",
-    });
+      const newAjaran = new TahunAjaran({ ajaran });
+
+      await newAjaran.save();
+
+      res.status(200).json({
+        success: true,
+        message: "Berhasil menambah Tahun Ajaran.",
+      });
+    } else {
+      throw new ResponseError(
+        404,
+        "Ajaran tidak valid. Pastikan perbedaan hanya 1 tahun."
+      );
+    }
   } catch (error) {
     next(error);
   }
@@ -85,13 +97,36 @@ export const editAjaran = async (req, res, next) => {
 export const deleteAjaran = async (req, res, next) => {
   try {
     const id = req.params.id;
+    const totalAjaran = await TahunAjaran.countDocuments();
+
+    if (totalAjaran === 1) {
+      throw new ResponseError(
+        404,
+        "Tidak bisa menghapus tahun ajaran. setidak ada 1 tahun ajaran yang aktif."
+      );
+    }
 
     const ajaran = await TahunAjaran.findByIdAndDelete(id);
 
     if (!ajaran) {
       throw new ResponseError(404, "Tahun Ajaran tidak ditemukan");
     }
-    console.log(ajaran);
+
+    if (ajaran.status) {
+      const newActiveAjaran = await TahunAjaran.findOneAndUpdate(
+        {},
+        { status: true },
+        { new: true, sort: { createdAt: 1 } }
+      );
+
+      if (!newActiveAjaran) {
+        throw new ResponseError(
+          500,
+          "Gagal mengaktifkan tahun ajaran lain setelah penghapusan."
+        );
+      }
+    }
+
     res.status(200).json({
       success: true,
       message: "Berhasil Menghapus Tahun Ajaran " + ajaran.ajaran,
